+++ src/sky.cpp
@@ -14,7 +14,7 @@ Sky::Sky(Game* parent) : Object(glm::vec3(),parent)
   editTextureCoord(0,0,0);
   for (int i = 0;i<5;i++)
   {
-    addPoint(i+1,glm::vec3(sin((63.43)/180*3.1415)*cos(i*3.1415/2.5)     ,cos((63.43)/180*3.1415),sin((63.43)/180*3.1415)*sin(i*3.1415/2.5))*radius,glm::vec3(0,1,0),1,1,1);
+    addPoint(i+1,glm::vec3((float)sin((63.43)/180*3.1415)*(float)cos(i*3.1415/2.5)     ,(float)cos((63.43)/180*3.1415),(float)sin((63.43)/180*3.1415)*(float)sin(i*3.1415/2.5))*radius,glm::vec3(0,1,0),1,1,1);
     editTextureCoord(i+1,i/5.f,i/10.f);
     addTriangle(i,0,i+1,(i+1)%5 +1);
   }
@@ -22,7 +22,7 @@ Sky::Sky(Game* parent) : Object(glm::vec3(),parent)
   editTextureCoord(6,0,0);
   for (int i = 0;i<5;i++)
   {
-    addPoint(i+7,glm::vec3(sin((63.43)/180*3.1415)*cos((i+0.5)*3.1415/2.5)     ,-cos((63.43)/180*3.1415),sin((63.43)/180*3.1415)*sin((i+0.5)*3.1415/2.5))*radius,glm::vec3(0,-1,0),1,1,1);
+    addPoint(i+7,glm::vec3((float)sin((63.43)/180*3.1415)*(float)cos((i+0.5)*3.1415/2.5)     ,-(float)cos((63.43)/180*3.1415),(float)sin((63.43)/180*3.1415)*(float)sin((i+0.5)*3.1415/2.5))*radius,glm::vec3(0,-1,0),1,1,1);
     editTextureCoord(i+7,i/5.f,i/10.f);
     addTriangle(i+5,6,i+7,(i+1)%5 + 7);
   }
@@ -36,7 +36,7 @@ Sky::Sky(Game* parent) : Object(glm::vec3(),parent)
   theta = 0.f;
 }
 
-inline float k(float x){return -cos(x*3.1415f)/2.f+0.5f;}
+inline float k(float x){return -(float)cos(x*3.1415f)/2.f+0.5f;}
  
 
 void Sky::Render(int refreshTime, glm::vec3 cameraPos)
@@ -51,8 +51,8 @@ void Sky::Render(int refreshTime, glm::vec3 cameraPos)
   // Update the colour...
   theta += refreshTime / 1000.f *3.1415f*2*2.f / 600.f;
   float l = pow(3.f*(1.f - k(k(k(theta/3.1415f))))/4.f + 0.25f,1.4f);
-  objectData.objectColour[0] = l*( k(k(k(sin(theta/2.f))))/20.f + 0.4f);
-  objectData.objectColour[1] = l*( 0.6f - k(k(k(sin(theta/2.f))))/9.f);
+  objectData.objectColour[0] = l*( k(k(k((float)sin(theta/2.f))))/20.f + 0.4f);
+  objectData.objectColour[1] = l*( 0.6f - k(k(k((float)sin(theta/2.f))))/9.f);
   objectData.objectColour[2] = l*1.f;
   updateObjectBO();
   // Render stuff
@@ -86,7 +86,7 @@ Sun::Sun(glm::vec3 pos,Game* parent) : Object(pos,parent)
   {
     addPoint(i,glm::vec3(2*SUN_SIZE*(float)sin(i*3.1415*2/20),500.f,2*SUN_SIZE*(float)cos(i*3.1415*2/20)),
       glm::vec3(0,1,0),1,1,1);
-    editTextureCoord(i,(float)(0.5+0.5*sin(i*3.1415*2/20)),(float)(0.5+0.5*cos(i*3.1415*2/20)));
+    editTextureCoord(i,(float)(0.5+0.5*(float)sin(i*3.1415*2/20)),(float)(0.5+0.5*(float)cos(i*3.1415*2/20)));
   }
   for(int i = 1;i<19;i++)
     addTriangle(i,0,i,i+1);
@@ -98,20 +98,20 @@ void Sun::Render(int refreshTime,glm::vec3 cameraPos)
   setPosition(cameraPos);
   theta += refreshTime / 1000.f *3.1415f*2*2.f / 600.f;
 
-  game->mainShader->frameData.sunDirection[0] = sin(theta);
-  game->mainShader->frameData.sunDirection[1] = cos(theta);
+  game->mainShader->frameData.sunDirection[0] = (float)sin(theta);
+  game->mainShader->frameData.sunDirection[1] = (float)cos(theta);
   game->mainShader->frameData.sunDirection[2] = 0;
 
-  rotate(glm::vec3(cos(theta),-sin(theta),0),
-         glm::vec3(sin(theta),cos(theta),0));
+  rotate(glm::vec3((float)cos(theta),-(float)sin(theta),0),
+         glm::vec3((float)sin(theta),(float)cos(theta),0));
   updateTriangleData();
   // We want to be fully lit.
   game->mainShader->frameData.sunIntensity=1.0;
   Object::Render(refreshTime,cameraPos);
   // Then light the rest as per usual.
   game->mainShader->frameData.sunIntensity = pow(3.f*(1.f - k(k(k(theta/3.1415f))))/4.f + 0.25f,1.4f);
-  if (cos(theta)>0)
-    game->mainShader->frameData.sunIntensity=cos(theta);
+  if ((float)cos(theta)>0)
+    game->mainShader->frameData.sunIntensity=(float)cos(theta);
   else
     game->mainShader->frameData.sunIntensity=0;
   game->mainShader->setFrameData();
@@ -132,7 +132,7 @@ Stars::Stars(Game* parent) : Object(glm::vec3(0),parent)
     for(int i = j*6;i<(j+1)*6;i++)
     {
       addPoint(i,po+up*(float)sin(i/6.0*3.1415*2)+(float)cos(i/6.0*3.1415*2)*right,glm::vec3(0,1,1),1,1,1);
-      editTextureCoord(i,(float)(0.5+0.5*sin(i*3.1415*2/6)),(float)(0.5+0.5*cos(i*3.1415*2/6)));
+      editTextureCoord(i,(float)(0.5+0.5*(float)sin(i*3.1415*2/6)),(float)(0.5+0.5*(float)cos(i*3.1415*2/6)));
     }
     for(int i = 0;i<5;i++)
       addTriangle(j*5+i,j*6,j*6+i,j*6+i+1);
@@ -150,7 +150,7 @@ void Stars::Render(int refreshTime,glm::vec3 cameraPos)
   glm::vec3 inc = glm::vec3(0,0,1.f);
   glm::vec3 right = glm::cross(glm::vec3(1.f,0,0),inc);
   glm::vec3 up = glm::cross(inc,right);
-  rotate(inc,up*sin(-theta)+right*cos(theta));
+  rotate(inc,up*(float)sin(-theta)+right*(float)cos(theta));
   float l = 3.f*(1.f - k(k(k(theta/3.1415f))))/4.f + 0.25f;
   objectData.objectColour[3] = k((1-l)*(1-l)*(1-l));
   updateObjectBO();
